Moved templates and added template to model
To support different templates for rendering the feed entries we have
added support for configuration of a default in settings.  To specify
the default template it should be specified as an accessible template
path under the property 'ckan.feeds.default.snippet'.

Each created feed can optionally specify the path to a different
template used for rendering but this file must already be in a template
folder that is already accessible to the CKAN instance.  If no template
is set then it will use the configured one, falling back to the one
defined in ckanext-feedcontent